requirements.txt
# Unit 1: Core Engine
pynput>=1.7
mss>=9.0
Pillow>=10.0
google-generativeai>=0.8
google-auth>=2.0
google-auth-oauthlib>=1.2

# Unit 2: UI & Configuration
customtkinter>=5.2
pyperclip>=1.8
tkinterweb>=3.24
markdown>=3.6

# Testing & Quality
pytest>=8.0
pytest-mock>=3.12
ruff>=0.4
